refactor(app): update context, pages, hooks, and library utils
