+++ packages/app/src/features/login/hooks/useLoginForm.ts
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { authenticationLoginRequestPublicSchema } from '../../../../../shared/schemas/auth.schema';
+import { authenticationLoginRequestPublicSchema } from '../../../../../shared/src/schemas/auth.schema';
 
 export const useLoginForm = () => {
   const [email, setEmail] = useState('');
