+++ packages/app/src/features/register/hooks/useRegisterForm.ts
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { authenticationRegisterRequestPublicSchema } from '../../../../../shared/schemas/auth.schema';
+import { authenticationRegisterRequestPublicSchema } from '../../../../../shared/src/schemas/auth.schema';
 
 export const useRegisterForm = () => {
   const [email, setEmail] = useState('');
