+++ packages/app/src/index.css
@@ -1,4 +1,24 @@
 @import 'tailwindcss';
+@font-face {
+  font-family: 'Inter';
+  src: url('/fonts/inter/Inter_VariableFont_opsz,wght.ttf') format('truetype');
+  font-weight: 100 900; /* supports all weights */
+  font-style: normal;
+  font-display: swap;
+}
+
+@font-face {
+  font-family: 'Inter';
+  src: url('/fonts/inter/Inter_Italic_VariableFont_opsz,wght.ttf')
+    format('truetype');
+  font-weight: 100 900;
+  font-style: italic;
+  font-display: swap;
+}
+
+body {
+  font-family: 'Inter', 'Helvetica Neue', Helvetica, Arial, sans-serif;
+}
 
 /* :root {
   font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
