+++ packages/app/src/pages/Dashboard.tsx
@@ -1,3 +1,4 @@
+import Todos from '../features/dashboard/Todos';
 import { useAuth } from '../hooks/useAuth';
 
 const DashboardPage = () => {
@@ -8,6 +9,7 @@ const DashboardPage = () => {
       <h1>Dashboard</h1>
       <p>Logged in as: {state.authUser?.email || 'No user'}</p>
       <p>User status: {state.authUser?.isActive ? 'Active' : 'Inactive'}</p>
+      <Todos />
     </div>
   );
 };
